Compile and run code in editor
