+++ .gitignore
@@ -19,4 +19,4 @@ bin
 obj
 
 # Output
-out
+build
