+++ src/Fable.REPL/Fable.REPL.fsproj
@@ -1,14 +1,33 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <TargetFramework>netstandard1.6</TargetFramework>
+    <DefineConstants>FABLE_COMPILER</DefineConstants>
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="../../../FSharp.Compiler.Service_fable/fcs/fcs-fable/fcs-fable.fsproj" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="../../../Fable/src/dotnet/Fable.Core/Util.fs" />
+    <Compile Include="../../../Fable/src/dotnet/Fable.Core/Compiler.fs" />
+    <Compile Include="../../../Fable/src/dotnet/Fable.Core/Fable.Core.fs" />
+    <Compile Include="../../../Fable/src/dotnet/Fable.Core/AST/AST.Common.fs" />
+    <Compile Include="../../../Fable/src/dotnet/Fable.Core/AST/AST.Fable.fs" />
+    <Compile Include="../../../Fable/src/dotnet/Fable.Core/AST/AST.Fable.Util.fs" />
+    <Compile Include="../../../Fable/src/dotnet/Fable.Core/AST/AST.Babel.fs" />
+    <Compile Include="../../../Fable/src/dotnet/Fable.Core/Plugins.fs" />
+    <Compile Include="../../../Fable/src/dotnet/Fable.Core/Import/Fable.Import.JS.fs" />
+    <Compile Include="../../../Fable/src/dotnet/Fable.Core/Fable.Core.JsInterop.fs" />
+    <Compile Include="../../../Fable/src/dotnet/Fable.Core/Fable.Core.Extensions.fs" />
+
+    <Compile Include="../../../Fable/src/dotnet/Fable.Compiler/Utils.fs"/>
+    <Compile Include="../../../Fable/src/dotnet/Fable.Compiler/Replacements.fs"/>
+    <Compile Include="../../../Fable/src/dotnet/Fable.Compiler/FSharp2Fable.Util.fs"/>
+    <Compile Include="../../../Fable/src/dotnet/Fable.Compiler/FSharp2Fable.fs"/>
+    <Compile Include="../../../Fable/src/dotnet/Fable.Compiler/Fable2Babel.fs"/>
+    <Compile Include="../../../Fable/src/dotnet/Fable.Compiler/State.fs"/>
+  
     <Compile Include="interfaces.fs" />
     <Compile Include="lexer.fs" />
     <Compile Include="fcs.fs" />
   </ItemGroup>
-  <Import Project="..\..\.paket\Paket.Restore.targets" />
 </Project>
