+++ src/Fable.REPL/fcs.fs
@@ -6,10 +6,18 @@ open Microsoft.FSharp.Compiler
 open Microsoft.FSharp.Compiler.SourceCodeServices
 open FsAutoComplete
 open Interfaces
+open Fable
+open Fable.AST
+open Fable.Core
+open Fable.State
 
-type Checker(c: InteractiveChecker) =
+type private CheckerImpl(c: InteractiveChecker) =
     member this.Checker = c
     interface IChecker
+
+type private CompilerImpl(c: Compiler) =
+    member this.Compiler = c
+    interface IFableCompiler
 
 type ParseResults =
     { ParseFile: FSharpParseFileResults
@@ -74,6 +82,9 @@ let convertGlyph glyph =
 let createChecker references readAllBytes =
     InteractiveChecker.Create(List.ofArray references, readAllBytes)
 
+let createCompiler replacements =
+    Compiler(replacements = replacements)
+
 let parseFSharpProject (checker: InteractiveChecker) fileName source =
     let parseResults, typeCheckResults, projectResults = checker.ParseAndCheckScript (fileName, source)      
     { ParseFile = parseResults
@@ -91,16 +102,54 @@ let getCompletionsAtLocation (parseResults: ParseResults) line col lineText = as
         { Name = decl.Name; Glyph = convertGlyph decl.Glyph })
 }
 
+let makeProjOptions (com: ICompiler) projFile =
+    let projOptions: FSharpProjectOptions =
+      { ProjectFileName = projFile
+        SourceFiles = [| |]
+        OtherOptions = [| |]
+        ReferencedProjects = [| |]
+        IsIncompleteTypeCheckEnvironment = false
+        UseScriptResolutionRules = false
+        LoadTime = DateTime.Now
+        UnresolvedReferences = None
+        OriginalLoadReferences = []
+        ExtraProjectInfo = None
+        Stamp = None }
+    projOptions
+
+let compileAst (com: Compiler) (parseResults: ParseResults) (fableCoreDir: string) fileName =
+    let fableCoreDir =
+        if fableCoreDir.StartsWith(".")
+        then FilePath fableCoreDir
+        else NonFilePath fableCoreDir
+    // let errors = com.ReadAllLogs() |> Map.tryFind "error"
+    // if errors.IsSome then failwith (errors.Value |> String.concat "\n")
+    let projectOptions = makeProjOptions com fileName
+    let project = Project(projectOptions, parseResults.CheckProject, fableCoreDir, isWatchCompile=false)
+    let file: Babel.Program =
+        FSharp2Fable.Compiler.transformFile com project project.CheckedProject fileName
+        |> Fable2Babel.Compiler.transformFile com project
+    let loc = defaultArg file.loc SourceLocation.Empty
+    Babel.Program(file.fileName, loc, file.body, file.directives, com.ReadAllLogs())
+
 type private ExportsImpl() =
     interface IFableREPL with
         member this.CreateChecker(references, readAllBytes) =
-            createChecker references readAllBytes |> Checker :> IChecker
+            createChecker references readAllBytes |> CheckerImpl :> IChecker
+        member this.CreateCompiler(replacements) =
+            let replacements = defaultArg replacements (upcast [||])
+            createCompiler replacements |> CompilerImpl :> IFableCompiler
         member this.ParseFSharpProject(checker, fileName, source) =
-            let c = checker :?> Checker
+            let c = checker :?> CheckerImpl
             parseFSharpProject c.Checker fileName source :> IParseResults
         member this.GetCompletionsAtLocation(parseResults:IParseResults, line:int, col:int, lineText:string) =
             let res = parseResults :?> ParseResults
             getCompletionsAtLocation res line col lineText
+        member this.CompileToBabelJsonAst(com: IFableCompiler, parseResults:IParseResults, fableCoreDir:string, fileName:string) =
+            let com = com :?> CompilerImpl
+            let res = parseResults :?> ParseResults
+            compileAst com.Compiler res fableCoreDir fileName |> JsInterop.toJson
+
 
 let Exports: IFableREPL = upcast ExportsImpl()
