--- src/Fable.REPL/paket.references
@@ -1,1 +1,0 @@
-Fable.Core
