+++ src/Monaco/util.js
@@ -1,5 +1,9 @@
 /// @ts-check
 
+import * as Babel from "babel-standalone";
+import BabelTemplate from "babel-template";
+import * as BabelPlugins from "../../../Fable/src/js/fable-utils/babel-plugins";
+
 var checker = null;
 var metadata = {}
 
@@ -26,18 +30,18 @@ function getFileBlob(key, url) {
     xhr.open("GET", url + ".txt");
     xhr.responseType = "arraybuffer";
     xhr.onload = function (oEvent) {
-      var arrayBuffer = xhr.response;
-      if (arrayBuffer) {
-        metadata[key] = new Uint8Array(arrayBuffer);
-      }
+        var arrayBuffer = xhr.response;
+        if (arrayBuffer) {
+            metadata[key] = new Uint8Array(arrayBuffer);
+        }
     };
     xhr.onerror = function (oEvent) {
-      console.log('Error loading ' + url);
+        console.log('Error loading ' + url);
     };
     xhr.send();
 };
 
-references.forEach(function(fileName){
+references.forEach(function (fileName) {
     getFileBlob(fileName, 'metadata/' + fileName);
 });
 
@@ -51,5 +55,37 @@ export function getChecker(createChecker) {
         checker = createChecker(references2, readAllBytes)
     }
     return checker;
+}
+
+export function runAst(jsonAst) {
+    try {
+        var options = {
+            plugins: [
+                BabelPlugins.getTransformMacroExpressions(BabelTemplate),
+                BabelPlugins.getRemoveUnneededNulls(),
+                // Use AMD modules for evaluation
+                "transform-es2015-modules-amd"
+            ],
+            filename: 'repl',
+            babelrc: false,
+        };
+
+        //   if (!document.getElementById('option-es2015').checked) {
+        //     options.presets = [
+        //       ["es2015", {"modules": false }]
+        //     ];
+        //   }
+
+        var ast = JSON.parse(jsonAst);
+        var code =
+            Babel.transformFromAst(ast, null, options).code
+                .replace("define", "require")
+                .replace('"use strict";', '"use strict"; try { exports = exports || {}; } catch (err) {}');
+
+        // Run
+        new Function(code)();
+    } catch (err) {
+        console.error(err.message + "\n" + err.stack);
+    }
 }
