--- src/js/babel-standalone.js
@@ -1,5 +1,0 @@
-var babel = require("babel-standalone");
-var template = require("babel-template");
-
-babel.template = template;
-module.exports = babel;
