--- src/js/repl.html
@@ -1,76 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" class="babel-fullscreen" class="no-js">
-<head>
-<meta charset="utf-8">
-<meta http-equiv="X-UA-Compatible" content="IE=edge">
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<title>Fable JS</title>
-<meta name="description" content="The compiler for writing next generation JavaScript">
-<meta name="viewport" content="width=device-width, initial-scale=1">
-<link rel="stylesheet" href="https://babeljs.io/css/main.css">
-<style>
-body { padding-top: 0; }
-.babel-repl-reporter { z-index: 100 }
-</style>
-
-</head>
-<div class="babel-repl">
-  <form class="form babel-repl-toolbar">
-    <div class="pull-left">
-      <!--<div class="form-group"> <label for="option-astInput"> <input id="option-astInput" type="checkbox" disabled> AST Input </label></div>
-      <div class="form-group"> <label for="option-evaluate"> <input id="option-evaluate" type="checkbox" disabled> Evaluate </label></div>-->
-      <div class="form-group button-group babel-repl-preset-container">
-      <span class="dropdown-toggle" data-toggle="dropdown"> Samples:
-        <span
-          class="caret"></span>
-          </span>
-          <ul class="dropdown-menu" id="sample-dropdown"></ul>
-      </div>
-      <!--<div class="form-group"> <label for="option-lineWrap"> <input id="option-lineWrap" type="checkbox"> Line Wrap </label></div>-->
-      <div class="form-group"> <label for="option-es2015"> <input id="option-es2015" type="checkbox" checked> ES2015</label></div>
-    </div>
-    <div class="pull-left">
-      <div class="form-group" style="color:black; padding-top: 10px">
-        <button type="button" class="btn btn-default" id="fabel-compile">Compile</button>
-        <button type="button" class="btn btn-default" id="fabel-run">Run</button>
-      </div>
-    </div>
-    <div class="pull-right">
-      <div class="form-group"> <strong>Fable v1.0.0-narumi, FCS v12.0.8, Babel v<span id="babel-repl-version"></span></strong></div>
-    </div>
-  </form>
-  <div class="babel-repl-left-panel">
-    <div class="babel-repl-editor babel-repl-input">
-      <div class="ace_editor"></div>
-    </div>
-    <div class="babel-repl-reporter babel-repl-errors"></div>
-  </div>
-  <div class="babel-repl-right-panel">
-    <div class="babel-repl-resize"></div>
-    <div class="babel-repl-editor babel-repl-output">
-      <div class="ace_editor"></div>
-    </div>
-    <div class="babel-repl-reporter babel-repl-console"></div>
-  </div>
-</div>
-<script src='https://cdnjs.cloudflare.com/ajax/libs/jquery/2.1.3/jquery.min.js' type="text/javascript"></script>
-<script src='https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.1/js/bootstrap.min.js' type="text/javascript"></script>
-<script src='https://cdnjs.cloudflare.com/ajax/libs/clipboard.js/1.5.13/clipboard.min.js' type="text/javascript"></script>
-<script src='https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.2/require.min.js' type="text/javascript"></script>
-<script src='https://babeljs.io/scripts/clipboard.js' type="text/javascript"></script>
-<script src='https://babeljs.io/scripts/open-in-repl.js' type="text/javascript"></script>
-<script>
-  $("html").removeClass("no-js").addClass("js");
-  requirejs.config({
-    paths: {
-      'fable-core': '/fable-core'
-    }
-  });
-</script>
-<script src='https://cdnjs.cloudflare.com/ajax/libs/ace/1.1.3/ace.js' type="text/javascript"></script>
-<script src="https://babeljs.io/scripts/pretty-format.js" type="text/javascript"></script>
-<script src='/repl/babel-standalone.js' type="text/javascript"></script>
-<script src='/repl_samples/samples.js' type="text/javascript"></script>
-<script src='/repl.js' type="text/javascript"></script>
-</body>
-</html>
